hotelbooking/pyproject.toml
[project]
name = "hh"
version = "1"
description = ""
authors = [
    {name = "Denis Martunets",email = "smind716@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "ruff (>=0.12.1,<0.13.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "types-pyyaml (>=6.0.12.20250516,<7.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
